Pulled changes from scratch org remote
The sfdx tracks and handles the changes made in the remote via a git like SCM tool under the hood.
This tool makes it possible to track changes made either in the scratch org remote or in the local development environment.
The overall SCM has been decoupled and left upto the developer to choose git or any other tool of their choice.
The sfdx cli provides built-in SCM to sync metadata or data changes between local and scratch org remote environments.